foodStories/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dateparser==0.7.6
defusedxml==0.7.0rc1
Django==3.0.6
django-3-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-rest-auth==0.9.5
django-taggit==1.3.0
django-tags-input==5.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
kombu==4.6.11
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.2
prompt-toolkit==2.0.9
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.17
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.4
zipp==3.1.0
